+++ CHANGES.rst
@@ -18,6 +18,7 @@ Improvements
 - Use a gender-neutral chairperson icon (:pr:`5710`)
 - Add option to set the abstracts' primary authors as the default submitters for the
   corresponding contributions (:pr:`5711`)
+- Allow commenting on accepted/rejected editables (:issue:`5712`, :pr:`5722`)
 
 Bugfixes
 ^^^^^^^^